Rename search-input to search-input-legacy
