+++ components/search-input-legacy/stats-search-item.js
